Iniciando a integracao com o AplicativoAPS
Acrescentou paginação na busca de ocorrências e alterou a busca
TipoOcorrencia para Lazy;
Acrescentou filtro de busca por Status;
Acrescentou um for na Classe PopulaBanco, para aumentar o número de
registro no banco;
Alterou o nome de algumas variáveis e corrigiu alguns erros de
digitação;

+++ APIREST_APS/src/main/java/br/com/aplicacaoaps/apirest/models/TipoOcorrencia.java
@@ -12,7 +12,6 @@ import javax.persistence.InheritanceType;
 public abstract class TipoOcorrencia {
 
 	@Id @GeneratedValue(strategy = GenerationType.AUTO)
-	protected int id;
+	protected Long id;
 
-	
 }
