+++ APIREST_APS/src/main/java/br/com/aplicacaoaps/apirest/models/TipoOcorrenciaRegional.java
@@ -26,7 +26,8 @@ public class TipoOcorrenciaRegional extends TipoOcorrencia {
 		this.tags = TagsRegional;
 	}
 
-	public int getId() {
+
+	public Long getId() {
 		return this.id;
 	}
